refactor reduce_scatter_program_factories
